fix log cache lock management snafu: * if we cannot delete the lock file, the cache is still in use   -> don't try to delete the cache file in that case
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@13223 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/LogCache/CachedLogInfo.cpp
@@ -79,8 +79,8 @@ void CCachedLogInfo::CCacheFileManager::AutoAcquire (const std::wstring& fileNam
     std::wstring lockFileName = fileName + L".lock";
     if (IsMarked (lockFileName))
     {
-        DeleteFile (fileName.c_str());
-        DeleteFile (lockFileName.c_str());
+        if (DeleteFile (lockFileName.c_str()) == TRUE)
+            DeleteFile (fileName.c_str());
     }
 
     // auto-create file and acquire lock
